+++ src/Common/NetworkCore/Network/Pipeline/Outbound/DeliveryMethodOutboundPipelineStage.cs
@@ -43,6 +43,7 @@ public class DeliveryMethodOutboundPipelineStage : IOutboundPipelineStage
         { EventId.Core_Movement_Move, DeliveryMethod.ReliableUnordered },
         { EventId.Core_Movement_RequestMove, DeliveryMethod.ReliableUnordered },
         { EventId.Core_Movement_Jump, DeliveryMethod.ReliableUnordered },
+        { EventId.Core_Movement_TeleportNotice, DeliveryMethod.ReliableUnordered },
         { EventId.Core_WorldManagement_EntityLeaveWorldSegment, DeliveryMethod.ReliableUnordered },
         { EventId.Core_Network_Logout, DeliveryMethod.ReliableUnordered },
         { EventId.Core_Chat_Send, DeliveryMethod.ReliableUnordered },